+++ webapp/tests/unit/utils/redis-config.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for Redis configuration validation (W-076)
  * @description     Tests Redis config parsing, validation, and fallback behavior
  * @file            webapp/tests/unit/utils/redis-config.test.js
- * @version         1.6.16
- * @release         2026-02-12
+ * @version         1.6.17
+ * @release         2026-02-14
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/